packages/k8s/pyproject.toml
#-------------------------------------------------------------------------------
# Project: k8s
#-------------------------------------------------------------------------------

[project]

name = "k8s"
version = "0.0.0"
requires-python = ">=3.12"

dependencies = [
    "aiohttp>=3.11.11",
    "anyio>=4.8.0",
    "pulumi>=3.147.0,<4.0.0",
    "pulumi-command>=1.0.1",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "tools",
]

#-------------------------------------------------------------------------------
# Build System
#-------------------------------------------------------------------------------

[build-system]

requires = ["hatchling"]
build-backend = "hatchling.build"

#-------------------------------------------------------------------------------
# UV
#-------------------------------------------------------------------------------

[tool.uv.sources]

tools = { workspace = true }
